+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+* Improve build system
+
 ## 0.1.0 - 2025-03-22
 
 Initial public release